+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.2] - 2021-09-26
 ### Added
 * The module version number in `webarchive.__version__`.
 * Initial support for command-line arguments in `extractor-gui.py`.
@@ -92,7 +92,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Initial public release.
 
-[Unreleased]: https://github.com/bmjcode/pywebarchive/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/bmjcode/pywebarchive/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/bmjcode/pywebarchive/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/bmjcode/pywebarchive/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/bmjcode/pywebarchive/compare/v0.2.4...v0.3.0
 [0.2.4]: https://github.com/bmjcode/pywebarchive/compare/v0.2.3...v0.2.4